\name{pythagorus-package}
\alias{pythagorus-package}
\alias{pythagorus}
\docType{package}
\title{
\packageTitle{pythagorus}
}
\description{
\packageDescription{pythagorus}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pythagorus}
\packageIndices{pythagorus}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pythagorus}

Maintainer: \packageMaintainer{pythagorus}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
